+++ memote/suite/cli/reports.py
@@ -22,7 +22,7 @@ from __future__ import absolute_import
 import logging
 import sys
 from builtins import open
-from multiprocessing import Pool
+from multiprocessing import Pool, cpu_count
 from functools import partial
 
 from libsbml import SBMLError
@@ -143,6 +143,7 @@ def snapshot(model, filename, pytest_args, exclusive, skip, solver,
                    "option can be specified multiple times.")
 def history(location, model, filename, deployment, custom_config):
     """Generate a report over a model's git commit history."""
+    LOGGER.info("Initialising history report generation.")
     if location is None:
         raise click.BadParameter("No 'location' given or configured.")
     try:
@@ -153,7 +154,9 @@ def history(location, model, filename, deployment, custom_config):
             "the model's commit history.")
         sys.exit(1)
     previous = repo.active_branch
-    repo.heads[deployment].checkout()
+    LOGGER.info("Obtaining history of results from "
+                "the deployment branch {}.".format(deployment))
+    repo.git.checkout(deployment)
     try:
         manager = managers.SQLResultManager(repository=repo, location=location)
     except (AttributeError, ArgumentError):
@@ -163,8 +166,10 @@ def history(location, model, filename, deployment, custom_config):
     # Update the default test configuration with custom ones (if any).
     for custom in custom_config:
         config.merge(ReportConfiguration.load(custom))
+    LOGGER.info("Tracing the commit history.")
     history = managers.HistoryManager(repository=repo, manager=manager)
     history.load_history(model, skip={deployment})
+    LOGGER.info("Composing the history report.")
     report = api.history_report(history, config=config)
     previous.checkout()
     with open(filename, "w", encoding="utf-8") as file_handle:
@@ -271,7 +276,7 @@ def diff(models, filename, pytest_args, exclusive, skip, solver,
     partial_test_diff = partial(_test_diff, pytest_args=pytest_args,
                                 skip=skip, exclusive=exclusive,
                                 experimental=experimental)
-    pool = Pool(min(len(models), os.cpu_count()))
+    pool = Pool(min(len(models), cpu_count()))
     results = pool.map(partial_test_diff, model_and_model_ver_tuple)
 
     for model_path, result in zip(models, results):