+++ tests/test_for_suite/conftest.py
@@ -17,13 +17,18 @@
 
 from __future__ import absolute_import
 
+import logging
+from os import chdir, getcwd
 from builtins import str
-from os.path import basename
+from os.path import basename, dirname, pardir, join
 from shutil import copyfile
-from git import Repo
+from git import Repo, InvalidGitRepositoryError
+from tarfile import open
 
 import pytest
 from click.testing import CliRunner
+
+LOGGER = logging.getLogger()
 
 
 @pytest.fixture(scope="session")
@@ -40,12 +45,35 @@ def model_file(small_file, tmpdir_factory):
 
 
 @pytest.fixture(scope="session")
+def invalid_file(invalid_model, tmpdir_factory):
+    filename = str(tmpdir_factory.mktemp("invalid_models").join(
+        basename(invalid_model)))
+    copyfile(invalid_model, filename)
+    return filename
+
+
+@pytest.fixture(scope="function")
 def mock_repo(tmpdir_factory):
     """
-    Clones the mock repo https://github.com/ChristianLieven/memote-mock-repo.
+    Unzips the mock repo providing the path to it and a repo instance.
 
     """
-    path = str(tmpdir_factory.mktemp("mock-repo"))
-    repo = Repo.clone_from(
-        'https://github.com/ChristianLieven/memote-mock-repo.git', path)
+    cwd = getcwd()
+    # Define the path to the temporary folder
+    base_path = str(tmpdir_factory.mktemp("mock-repo"))
+    # and the tarball that we will decompress there
+    tar_file = join(
+        dirname(__file__), pardir, "data", "memote-mock-repo.tar.gz"
+    )
+    # Extract the gzipped tarball
+    mock_repo_tarfile = open(tar_file)
+    mock_repo_tarfile.extractall(base_path)
+    mock_repo_tarfile.close()
+    # Obtain the repository as a gitpython Repo object
+    path = join(base_path, "memote-mock-repo")
+    try:
+        repo = Repo(path)
+    except InvalidGitRepositoryError:
+        LOGGER.warning(
+            "Could not find memote-mock-repository. Is the path correct?")
     return path, repo